langchainServeFlask/requirements.txt
llama_index
langchain
openai
python-dotenv
flask
psycopg2-binary